Move right hand rule to external file
