+++ object.js
@@ -1,3 +1,5 @@
+var rightHandRule = require('./rhr');
+
 /**
  * @alias geojsonhint
  * @param {(string|object)} GeoJSON given as a string or as an object
@@ -221,52 +223,6 @@ function hint(gj, options) {
         }
     }
 
-    function rightHandRule (geometry) {
-        var rhr = true;
-        if (geometry.type == 'Polygon') {
-            rhr = isPolyRHR(geometry.coordinates);
-        } else if (geometry.type == 'MultiPolygon') {
-            if (!geometry.coordinates.every(isPolyRHR))
-                rhr = false;
-        }
-        if (!rhr) {
-            errors.push({
-                message: 'Polygons and MultiPolygons should follow the right-hand rule',
-                level: 'warn',
-                line: geometry.__line__
-            });
-        }
-    }
-
-    function isPolyRHR (coords) {
-        if (coords && coords.length > 0) {
-            if (!isRingClockwise(coords[0]))
-                return false;
-            var interiorCoords = coords.slice(1, coords.length);
-            if (interiorCoords.some(isRingClockwise))
-                return false;
-        }
-        return true;
-    }
-
-    function isRingClockwise (coords) {
-        var area = 0;
-        if (coords.length > 2) {
-            var p1, p2;
-            for (var i = 0; i < coords.length - 1; i++) {
-                p1 = coords[i];
-                p2 = coords[i + 1];
-                area += rad(p2[0] - p1[0]) * (2 + Math.sin(rad(p1[1])) + Math.sin(rad(p2[1])));
-            }
-        }
-
-        return area >= 0;
-    }
-
-    function rad(x) {
-        return x * Math.PI / 180;
-    }
-
     function crs(_) {
         if (!_.crs) return;
         var defaultCRSName = "urn:ogc:def:crs:OGC:1.3:CRS84";
@@ -346,7 +302,7 @@ function hint(gj, options) {
         bbox(polygon);
         if (!requiredProperty(polygon, 'coordinates', 'array')) {
             if (!positionArray(polygon.coordinates, 'LinearRing', 2)) {
-                rightHandRule(polygon);
+                rightHandRule(polygon, errors);
             }
         }
     }
@@ -357,7 +313,7 @@ function hint(gj, options) {
         bbox(multiPolygon);
         if (!requiredProperty(multiPolygon, 'coordinates', 'array')) {
             if (!positionArray(multiPolygon.coordinates, 'LinearRing', 3)) {
-                rightHandRule(multiPolygon);
+                rightHandRule(multiPolygon, errors);
             }
         }
     }
